refactor: shifting store_api_call to model level
