+++ integrated_channels/integrated_channel/models.py
@@ -925,3 +925,40 @@ class IntegratedChannelAPIRequestLogs(TimeStampedModel):
             f", response_body: {self.response_body}"
             f", status_code: {self.status_code}"
         )
+
+    @classmethod
+    def store_api_call(
+        cls,
+        enterprise_customer,
+        enterprise_customer_configuration_id,
+        endpoint,
+        payload,
+        time_taken,
+        status_code,
+        response_body,
+    ):
+        """
+        Creates new record in IntegratedChannelAPIRequestLogs table.
+        """
+        try:
+            record = cls(
+                enterprise_customer=enterprise_customer,
+                enterprise_customer_configuration_id=enterprise_customer_configuration_id,
+                endpoint=endpoint,
+                payload=payload,
+                time_taken=time_taken,
+                status_code=status_code,
+                response_body=response_body,
+            )
+            record.save()
+        except Exception as e:  # pylint: disable=broad-except
+            LOGGER.error(
+                f"store_api_call raised error while storing API call: {e}"
+                f"enterprise_customer={enterprise_customer}"
+                f"enterprise_customer_configuration_id={enterprise_customer_configuration_id},"
+                f"endpoint={endpoint}"
+                f"payload={payload}"
+                f"time_taken={time_taken}"
+                f"status_code={status_code}"
+                f"response_body={response_body}"
+            )
